🚚 [rendering] Move module `adapters.cookiecutter.{render => }`
