+++ src/cutty/rendering/adapters/cookiecutter.py
